backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pymysql==1.1.1
cryptography==43.0.1
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.1
scipy==1.14.1
